Small refactoring to address PR comments
