+++ examples/monitor/match_bets_with_langfuse_traces.py
@@ -4,7 +4,6 @@ from typing import Any
 import dotenv
 from eth_typing import HexAddress, HexStr
 
-from examples.monitor.financial_metrics import SharpeRatioCalculator
 from examples.monitor.transaction_cache import TransactionBlockCache
 from prediction_market_agent_tooling.markets.omen.omen_contracts import (
     OmenConditionalTokenContract,
@@ -12,6 +11,7 @@ from prediction_market_agent_tooling.markets.omen.omen_contracts import (
 from prediction_market_agent_tooling.markets.omen.omen_subgraph_handler import (
     OmenSubgraphHandler,
 )
+from prediction_market_agent_tooling.monitor.financial_metrics.financial_metrics import SharpeRatioCalculator
 
 dotenv.load_dotenv()
 import pandas as pd
